Update Authoring_quick_start_2.md
Line 43. Removed negative from \(\frac{3}{(x-1)^{4}}\),  since power is now in the denominator!

Question: Line 32.  I note that students now cannot enter either single or multi-character variable names that are defined in the question within their answer. If this is the new intention (fine by me!) then this line is not correct.